move launch code outside of main lib
