+++ cli/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-loki = { path = "../"}
+launch = { path = "../launch"}
 structopt = "0.3"
 serde = "1.0"
 serde_json = "1.0"
