+++ launch/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "launch"
+version = "0.1.0"
+authors = ["Pascal Benchimol <pascal.benchimol@kisio.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+loki = { path = "../"}
+serde = { version = "1.0", features = ["derive"] }
