+++ launch/src/config/input_data_type.rs
