+++ src/traits.rs
@@ -35,7 +35,7 @@
 // www.navitia.io
 
 use crate::{
-    config,
+    // config,
     loads_data::{Load, LoadsData},
     response,
     time::{PositiveDuration, SecondsSinceDatasetUTCStart},
@@ -407,7 +407,10 @@ where
     pub departure_datetime: NaiveDateTime,
     pub departures_stop_point_and_fallback_duration: Departures,
     pub arrivals_stop_point_and_fallback_duration: Arrivals,
-    pub params: config::RequestParams,
+    pub leg_arrival_penalty: PositiveDuration,
+    pub leg_walking_penalty: PositiveDuration,
+    pub max_nb_of_legs: u8,
+    pub max_journey_duration: PositiveDuration,
 }
 
 pub trait RequestIO<'data, Data: self::Data>: Request {
@@ -495,21 +498,4 @@ impl fmt::Display for BadRequest {
     }
 }
 
-pub trait Solver<'data, Data> {
-    fn new(nb_of_stops: usize, nb_of_missions: usize) -> Self;
 
-    fn solve_request<Departures, Arrivals, D, A>(
-        &mut self,
-        data: &'data Data,
-        model: &transit_model::Model,
-        request_input: RequestInput<Departures, Arrivals, D, A>,
-        comparator: &config::ComparatorType,
-    ) -> Result<Vec<response::Response>, BadRequest>
-    where
-        Self: Sized,
-        Arrivals: Iterator<Item = (A, PositiveDuration)>,
-        Departures: Iterator<Item = (D, PositiveDuration)>,
-        A: AsRef<str>,
-        D: AsRef<str>,
-        Data: self::DataWithIters;
-}
